// generated/api/gameDetailed.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * juniors bootcamp backend 🔥
 * Апи для выполнения индивидуальных заданий
 * OpenAPI spec version: 1.0
 */
import type { GameDeliveryType } from "./gameDeliveryType";
import type { GameGenre } from "./gameGenre";
import type { GameSystemRequirements } from "./gameSystemRequirements";
import type { GameType } from "./gameType";

export interface GameDetailed {
  /** Slug игры */
  slug: string;
  /** Название игры */
  name: string;
  /** Дата релиза */
  releaseDate: number;
  /** Тип */
  type: GameType;
  /** Жанры игры */
  genres: GameGenre[];
  /** Изображение игры */
  image: string;
  /** Способ получения */
  deliveryTypes: GameDeliveryType[];
  /** Описание игры */
  description: string;
  /** Минимальные системные требования */
  minimumSystemRequirements: GameSystemRequirements;
  /** Рекомендуемые системные требования */
  recommendedSystemRequirements: GameSystemRequirements;
  /** Разработчик */
  developer: string;
  /** Издатель */
  publisher: string;
  /** Внешний ID (Steam/KupiKod) */
  externalId: string;
  /** Скриншоты игры */
  screenshots: string[];
}